cn: introduce PeerSet interface to test ProtocolManager
